+++ src/motion.rs
@@ -2,7 +2,7 @@ use crate::Duration;
 use crate::animations::spring::{Spring, SpringState};
 use crate::animations::utils::{Animatable, AnimationMode};
 use crate::keyframes::KeyframeAnimation;
-use crate::prelude::{AnimationConfig, LoopMode, Transform, Tween};
+use crate::prelude::{AnimationConfig, LoopMode, Tween};
 use crate::sequence::AnimationSequence;
 use std::sync::Arc;